# [Enhancement] optimize decode_string for DecodeNode